+++ testing-bootstrap.js
@@ -23,9 +23,13 @@ require('zone.js/dist/async-test');
 Error.stackTraceLimit = Infinity;
 var testing = require('@angular/core/testing');
 var browser = require('@angular/platform-browser-dynamic/testing');
+var forms = require('@angular/forms');
+
 testing.setBaseTestProviders(
     browser.TEST_BROWSER_DYNAMIC_PLATFORM_PROVIDERS,
-    browser.TEST_BROWSER_DYNAMIC_APPLICATION_PROVIDERS
+    browser.TEST_BROWSER_DYNAMIC_APPLICATION_PROVIDERS,
+    forms.disableDeprecatedForms(),
+    forms.provideForms()
 );
 
 require('./src/index.ts');